tests/requirements.txt
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
fairlearn==0.11.0
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
patra-toolkit==0.1.2
pillow==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
shap==0.46.0
six==1.17.0
slicer==0.0.8
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
